+++ test/geos_capi/geometry_collection_test.rb
@@ -34,4 +34,15 @@ class GeosGeometryCollectionTest < Minitest::Test # :nodoc:
     assert_equal(noded.count, 4)
     assert(expected_lines.all? { |line| noded.include? line })
   end
+
+  def test_polygonize_collection
+    input = @factory.parse_wkt(
+      "GEOMETRYCOLLECTION(LINESTRING(0 0, 1 1, 1 0, 0 0), POINT(2 2))"
+    )
+    expected = @factory.parse_wkt(
+      "GEOMETRYCOLLECTION(POLYGON ((0 0, 1 1, 1 0, 0 0)))"
+    )
+
+    assert_equal expected, input.polygonize
+  end
 end if RGeo::Geos.capi_supported?